test(inference): Story 08 full-suite stabilization — sweep paid, zero branch-new failures
The owed full-suite sweep (isolated HOME, -n auto, Python 3.13.11): 148
failures triaged against a fresh main baseline (72 = the known inherited
local-env set) into 70 inherited + 78 branch-new. All 78 classified
a/b/c and repaired:

- (b) no-reservation replay read a controller receipt with the narrow
  SERVICE principal → fallback receipt authority
  (inference_adoption_service.py)
- (b) C1 bound executor bypassed the intel.model_unavailable fault seam
  and collapsed the named fault out of retry evidence (intel_queue.py)
- (b) meeting modules over the 600-line density budget → routed-child /
  bound-deferred code extracted to meeting_session/intel_routed_children.py
  and meeting_session/deferred_bound.py
- (c) suite-wide pollution: test_phase143_inference_assignments._profile
  probed host llama_cpp and blanket-updated every readiness observation
  row (29 downstream 409s) → readiness pinned by observation id + leak
  proof test
- (a) old-posture rigs updated to ruled law (explicit live/speech
  bundles, immutable ij_ queue ids, frozen language/backend triples,
  startup speech-migration revision deltas); streamed intel_token
  contract retirement completed across Python and web registries

Final quiet-tree sweep: 6392 passed, 71 failed, zero branch-new — every
failure name reproduces on main (one inherited deadline test is a load
flake, serial-green twice). Full record: verification ledger §Full-suite
stabilization sweep. Next: the C1 counsel checkpoint.

PMO-Story: HS-143-08
PMO-Contract-Digest: sha256:910a68b5ba3d7c026fb05b41e2fe6ebebf52ae671396c867c3d1ed2d05b07cff
