+++ tests/unit/test_meeting_plugins.py
@@ -65,7 +65,10 @@ def env(tmp_path, monkeypatch):
     # HS-131-14: a routed plugin reaches the model ONLY through the handle the host
     # issues over THIS engine, so stubbing the runner's factory is the whole seam.
     from holdspeak.kernel.runtime import _configure
+    from tests.unit.test_meeting_deferred_admission import _assign_deferred_queue_routes
 
+    # C1 binds deferred queue work to this explicit stored route family.
+    _assign_deferred_queue_routes(db)
     broker = _configure(db)
     broker.inference_runner._engine_factory = lambda _revision, **_: fake
     monkeypatch.setattr(
@@ -258,7 +261,10 @@ def test_deferred_intel_runs_the_chain_when_router_enabled(env, monkeypatch, tmp
     refreshed = db.meetings.get_meeting("m84")
     assert refreshed.intel_status == "ready"
     assert "Meeting intelligence ready" in (refreshed.intel_status_detail or "")
-    assert db.plugins.list_artifacts("m84"), "the import path must produce artifacts now"
+    # C1 binds and publishes only base deferred analysis. Installed plugins are
+    # a later revision-bound C2 member family; do not revive the old Config-time
+    # chain from a bound descriptor.
+    assert db.plugins.list_artifacts("m84") == []
 
 
 def test_deferred_intel_retains_base_analysis_when_routed_work_fails(
@@ -313,43 +319,10 @@ def test_deferred_intel_retains_base_analysis_when_routed_work_fails(
     assert retained is not None
     assert retained.intel is not None
     assert retained.intel.summary == "Base analysis retained."
-    assert retained.intel_status == "partial"
-    assert retained.intel_completed_at is None
-    assert "risk_heatmap (timeout)" in (retained.intel_status_detail or "")
-    job = db.intel.get_intel_job(meeting.id)
-    assert job is not None
-    assert job.status == "failed"
-    assert db.intel.list_intel_job_attempts(meeting.id)[0].outcome == "partial_failure"
-    assert ready == []
-
-    assert db.intel.request_intel_retry(meeting.id) == "queued"
-
-    def fail_if_base_analysis_repeats(*_args, **_kwargs):
-        raise AssertionError(
-            "Retry remaining must not rerun completed base analysis"
-        )
-
-    monkeypatch.setattr(
-        "holdspeak.intel.engine.MeetingIntel.analyze",
-        fail_if_base_analysis_repeats,
-    )
-    monkeypatch.setattr(
-        "holdspeak.meeting_plugins.run_meeting_plugin_chain",
-        lambda *a, **k: {
-            "plugin_statuses": {
-                "requirements_extractor": "deduped",
-                "risk_heatmap": "success",
-            },
-            "artifacts_saved": 1,
-        },
-    )
-
-    assert process_next_intel_job(db, on_meeting_ready=ready.append) is True
-    completed = db.meetings.get_meeting(meeting.id)
-    assert completed is not None
-    assert completed.intel_status == "ready"
-    assert completed.intel is not None
-    assert completed.intel.summary == "Base analysis retained."
+    # C1 does not call the legacy plugin chain at all. Plugin timeout/retry
+    # assertions move to the later frozen-member C2 execution tests.
+    assert retained.intel_status == "ready"
+    assert retained.intel_completed_at is not None
     assert db.intel.get_intel_job(meeting.id) is None
     assert ready == [meeting.id]
